Bases at half price to build and half price to keep
  AIRLINE_BASE_COST_PERCENT=50
  AIRLINE_BASE_UPKEEP_PERCENT=50

Both scale a curve that multiplies by 1.7 per level, so the top levels are
where the money actually is - a level 10 base costs about forty times a level
3. Halving is the difference between a second base being a project and being
out of reach on a server where nobody grinds for hours a day.

Measured at Lisbon after the change:

  level  1       5,554,799 to build      10,318 a week
  level  3      16,053,369              29,820
  level  5      46,394,237              86,181
  level  8     227,934,889             423,407
  level 10     658,731,829           1,223,649

Percentages rather than a hardcoded half, so the next adjustment is a number
in game-settings.env; 100 is the game as it ships.

+++ airline-data/src/main/scala/com/patson/model/AirlineBase.scala
@@ -13,17 +13,19 @@ case class AirlineBase(airline : Airline, airport : Airport, countryCode : Strin
       0
     } else {
       var baseCost = (1000000 + airport.rating.overallRating * 120000).toLong
-      (baseCost * airportSizeRatio * Math.pow (COST_EXPONENTIAL_BASE, (scale - 1) )).toLong
+      (baseCost * airportSizeRatio * Math.pow (COST_EXPONENTIAL_BASE, (scale - 1) )
+        * GameConfig.baseCostPercent / 100).toLong
     }
   }
 
   val COST_EXPONENTIAL_BASE = 1.7
-  
+
   lazy val getUpkeep : Long = {
     val adjustedScale = if (scale == 0) 1 else scale //for non-existing base, calculate as if the base is 1
     var baseUpkeep = (5000 + airport.rating.overallRating * 150).toLong
 
-    (baseUpkeep * airportSizeRatio * Math.pow(COST_EXPONENTIAL_BASE, adjustedScale - 1)).toInt
+    (baseUpkeep * airportSizeRatio * Math.pow(COST_EXPONENTIAL_BASE, adjustedScale - 1)
+      * GameConfig.baseUpkeepPercent / 100).toInt
   }
 
   lazy val airportSizeRatio =
